Split filters and marshalling methods into own files
